fdfs

+++ Tables/DimTargetLevel.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[DimTargetLevel]
 (
 	CompanyKey tinyint,
-	TargetLevelKey  tinyint   NOT NULL ,
+	TargetLevelKey  int   NOT NULL ,
 	DgreeName  Nvarchar(100) not null ,
 	UpdatLasteTime datetime default getdate(),
 CONSTRAINT [PK_DimTargetLevel_1] PRIMARY KEY CLUSTERED
